+++ Source/JNA/waffle-jna-jakarta/src/test/java/waffle/windows/auth/WindowsSecurityContextTests.java
@@ -49,7 +49,7 @@ public class WindowsSecurityContextTests {
                 WindowsAccountImpl.getCurrentUsername());
         Assertions.assertTrue(ctx.isContinue());
         Assertions.assertEquals(securityPackage, ctx.getSecurityPackage());
-        assertThat(ctx.getToken().length).isGreaterThan(0);
+        assertThat(ctx.getToken().length).isPositive();
         ctx.dispose();
     }
 }