+++ Source/JNA/waffle-tests/src/main/java/waffle/mock/MockFailedAuthenticationWindowsAuthProvider.java
@@ -21,17 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-/*
-  Waffle (https://github.com/Waffle/waffle)
-
-  Copyright (c) 2010-2019 Application Security, Inc.
-
-  All rights reserved. This program and the accompanying materials are made available under the terms of the Eclipse
-  Public License v1.0 which accompanies this distribution, and is available at
-  https://www.eclipse.org/legal/epl-v10.html.
-
-  Contributors: Application Security, Inc.
- */
 package waffle.mock;
 
 import java.nio.charset.StandardCharsets;
@@ -46,9 +35,6 @@ import waffle.windows.auth.*;
  * @author dblock[at]dblock[dot]org
  */
 public class MockFailedAuthenticationWindowsAuthProvider implements IWindowsAuthProvider {
-
-    /** The Constant GUEST. */
-    private static final String GUEST = "Guest";
 
     /** The groups. */
     private final List<String> groups = new ArrayList<>();