+++ Source/JNA/waffle-tomcat85/src/test/java/waffle/apache/NegotiateAuthenticatorTests.java
@@ -209,7 +209,7 @@ public class NegotiateAuthenticatorTests {
                     final GenericWindowsPrincipal windowsPrincipal = (GenericWindowsPrincipal) request
                             .getUserPrincipal();
                     Assertions.assertTrue(windowsPrincipal.getSidString().startsWith("S-"));
-                    assertThat(windowsPrincipal.getSid().length).isGreaterThan(0);
+                    assertThat(windowsPrincipal.getSid().length).isPositive();
                     Assertions.assertTrue(windowsPrincipal.getGroups().containsKey("Everyone"));
                     assertThat(response.getHeaderNames().size()).isLessThanOrEqualTo(1);
                     break;
@@ -222,7 +222,7 @@ public class NegotiateAuthenticatorTests {
                 final String continueToken = response.getHeader("WWW-Authenticate")
                         .substring(securityPackage.length() + 1);
                 final byte[] continueTokenBytes = Base64.getDecoder().decode(continueToken);
-                assertThat(continueTokenBytes.length).isGreaterThan(0);
+                assertThat(continueTokenBytes.length).isPositive();
                 final ManagedSecBufferDesc continueTokenBuffer = new ManagedSecBufferDesc(Sspi.SECBUFFER_TOKEN,
                         continueTokenBytes);
                 clientContext.initialize(clientContext.getHandle(), continueTokenBuffer,
@@ -292,7 +292,7 @@ public class NegotiateAuthenticatorTests {
                 Assertions.assertEquals(401, response.getStatus());
                 continueToken = response.getHeader("WWW-Authenticate").substring(securityPackage.length() + 1);
                 continueTokenBytes = Base64.getDecoder().decode(continueToken);
-                assertThat(continueTokenBytes.length).isGreaterThan(0);
+                assertThat(continueTokenBytes.length).isPositive();
                 continueTokenBuffer = new ManagedSecBufferDesc(Sspi.SECBUFFER_TOKEN, continueTokenBytes);
                 clientContext.initialize(clientContext.getHandle(), continueTokenBuffer,
                         WindowsAccountImpl.getCurrentUsername());